feat: library refactor/reorganization, split fns into separate files
